Add test for passing non-trivial structures as varargs
Fixes #9042.
